Add CLI options and continuously scan patch dir
The patch server was so far fixed to run on port 80 which would require
elevated privileges. We obviously don't want to force this on the user
and thus now provide a command line option to set the port to something
else. Similarly, the directory where the patches are can also be
adjusted.

Additionally, to avoid having to restart the server when a new patch
arrives, it will now automatically rescan the patch directory when a
change is detected.